packaging: @openibex/core and tsup setup
- Bundling via tsup
- node-datachannel package required by helia/webrtp is excluded from
  packaging (due to being broken for packaging)
- Minor bugfixes / refactoring
