+++ package.json
@@ -42,6 +42,7 @@
     "lerna": "^8.1.3",
     "rimraf": "^5.0.7",
     "ts-jest": "^29.1.4",
+    "tsup": "^8.3.0",
     "tsx": "^4.11.2"
   }
 }
