+++ packages/core/README.md
@@ -1,0 +1,2 @@
+# @openibex/core
+Description of the module TODO
