+++ packages/core/src/core.ts
@@ -16,6 +16,20 @@ export interface OiLoggerInterface {
   // Add other methods if needed
 }
 
+export type OiValueSchema = { 
+  datatype: string, 
+  value: string 
+};
+
+export enum OiValueType {
+  String = "string",
+  Number = "number",
+  Boolean = "boolean",
+  BigInt = "bigint",
+  Symbol = "symbol",
+  Null = "null",
+  Undefined = "undefined"
+}
 
 // OiCore-Singleton
 let core: OiCore | undefined = undefined;
