--- packages/core/src/index.d.ts
@@ -1,1 +1,0 @@
-declare module '@openibex/core';
